+++ frontend/public/CHANGELOG.md
@@ -1,3 +1,15 @@
+## [0.9.25] - 2025-09-25
+
+### Added
+
+- [#96](https://github.com/mediux-team/AURA/issues/96) - Added new popup confirmation for destructive actions
+
+### Fixed
+
+- Fixed issue with clearing app cache not working properly on User Preferences Store
+
+---
+
 ## [0.9.24] - 2025-09-25
 
 ### Added